+++ stac_check/cli.py
@@ -1,5 +1,6 @@
+import importlib.metadata
+
 import click
-import pkg_resources
 
 from .lint import Linter
 from .logo import logo
@@ -38,7 +39,9 @@ def recursive_message(linter: Linter) -> None:
     click.secho(f"Max-depth = {linter.max_depth}")
     click.secho("-------------------------")
     for count, msg in enumerate(linter.validate_all):
-        click.secho(f"Asset {count+1} Validated: {msg['path']}", bg="white", fg="black")
+        click.secho(
+            f"Asset {count + 1} Validated: {msg['path']}", bg="white", fg="black"
+        )
         click.secho()
         if msg["valid_stac"] == True:
             recursive_linter = Linter(msg["path"], recursive=True)
@@ -177,7 +180,7 @@ def cli_message(linter: Linter) -> None:
 )
 @click.command()
 @click.argument("file")
-@click.version_option(version=pkg_resources.require("stac-check")[0].version)
+@click.version_option(version=importlib.metadata.distribution("stac-check").version)
 def main(file, recursive, max_depth, assets, links):
     linter = Linter(
         file, assets=assets, links=links, recursive=recursive, max_depth=max_depth