+++ stac_check/lint.py
@@ -1,9 +1,10 @@
+import importlib.metadata
+import importlib.resources
 import json
 import os
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Union
 
-import pkg_resources
 import requests
 import yaml
 from dotenv import load_dotenv
@@ -134,7 +135,9 @@ class Linter:
             self.message["asset_type"] if "asset_type" in self.message else ""
         )
         self.version = self.message["version"] if "version" in self.message else ""
-        self.validator_version = pkg_resources.require("stac-validator")[0].version
+        self.validator_version = importlib.metadata.distribution(
+            "stac-validator"
+        ).version
         self.validate_all = self.recursive_validation(self.item)
         self.valid_stac = self.message["valid_stac"]
         self.error_type = self.check_error_type()
@@ -185,7 +188,9 @@ class Linter:
             with open(default_config_file) as f:
                 default_config = yaml.load(f, Loader=yaml.FullLoader)
         else:
-            with pkg_resources.resource_stream(__name__, "stac-check.config.yml") as f:
+            with importlib.resources.open_text(
+                "stac_check", "stac-check.config.yml"
+            ) as f:
                 default_config = yaml.load(f, Loader=yaml.FullLoader)
         if config_file:
             with open(config_file) as f: