+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 
 ### Changed
 
+## [v1.5.0] - 2025-01-17
+
+### Changed
+
 - No longer use the deprecated pkg-resources package.
   It has been replaced with importlib from the Python standard library
   ([#112](https://github.com/stac-utils/stac-check/pull/112))
@@ -158,7 +162,8 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 - Validation from stac-validator 2.3.0
 - Links and assets validation checks
 
-[Unreleased]: https://github.com/stac-utils/stac-check/compare/v1.4.0...main
+[Unreleased]: https://github.com/stac-utils/stac-check/compare/v1.5.0...main
+[v1.5.0]: https://github.com/stac-utils/stac-check/compare/v1.4.0...v1.5.0
 [v1.4.0]: https://github.com/stac-utils/stac-check/compare/v1.3.3...v1.4.0
 [v1.3.3]: https://github.com/stac-utils/stac-check/compare/v1.3.2...v1.3.3
 [v1.3.2]: https://github.com/stac-utils/stac-check/compare/v1.3.1...v1.3.2